move to packages to be able to publish as npm pack
