+++ packages/functor/tsconfig.json
@@ -12,13 +12,13 @@
     // "jsx": "preserve",                     /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
     "declaration": true,                      /* Generates corresponding '.d.ts' file. */
     "declarationMap": true,                   /* Generates a sourcemap for each corresponding '.d.ts' file. */
-    "sourceMap": true,                        /* Generates corresponding '.map' file. */
     // "outFile": "./",                       /* Concatenate and emit output to single file. */
     "declarationDir": "../out/functor/typings",
+
     "outDir": "../out/functor",               /* Redirect output structure to the directory. */
     "rootDir": "./",                          /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     // "composite": true,                     /* Enable project compilation */
-    "tsBuildInfoFile": "../out/tmp",           /* Specify file to store incremental compilation information */
+    "tsBuildInfoFile": "../out/tmp/functor.log",           /* Specify file to store incremental compilation information */
     "removeComments": true,                   /* Do not emit comments to output. */
     "noEmit": false,                          /* Do not emit outputs. */
     "importHelpers": true,                    /* Import emit helpers from 'tslib'. */
@@ -56,8 +56,8 @@
     /* Source Map Options */
     // "sourceRoot": "",                      /* Specify the location where debugger should locate TypeScript files instead of source locations. */
     // "mapRoot": "",                         /* Specify the location where debugger should locate map files instead of generated locations. */
-    // "inlineSourceMap": true,               /* Emit a single file with source maps instead of having a separate file. */
-    // "inlineSources": true,                 /* Emit the source alongside the sourcemaps within a single file; requires '--inlineSourceMap' or '--sourceMap' to be set. */
+    "inlineSourceMap": true,               /* Emit a single file with source maps instead of having a separate file. */
+    "inlineSources": true,                 /* Emit the source alongside the sourcemaps within a single file; requires '--inlineSourceMap' or '--sourceMap' to be set. */
 
     /* Experimental Options */
     // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
